Rename fetch_artist_mbids.py
Rename fetch_artist_mbids.py to fetch_and_store_artist_mbids.py as
fetch_artist_mbids doesnot give user clue about the fact that this
script also stores the fetched MBIDs.
