+++ manage.py
@@ -1,6 +1,6 @@
 from messybrainz import db
-from messybrainz.fetch_artist_mbids import truncate_recording_artist_join,\
-                                        fetch_and_store_artist_mbids_for_all_recording_mbids
+from messybrainz.fetch_and_store_artist_mbids import truncate_recording_artist_join,\
+                                                fetch_and_store_artist_mbids_for_all_recording_mbids
 from messybrainz.webserver import create_app
 from brainzutils import musicbrainz_db
 from sqlalchemy import text
